+++ ChangeLog.md
@@ -1,9 +1,9 @@
 # Changelog
 
 ## Version [0.10.0](https://github.com/cedx/html.net/compare/v0.9.0...v0.10.0)
-- Added the `New-QElement`, `New-SlotElement` and `New-SourceElement` cmdlets.
-- Added the `q`, `slot` and `source` aliases.
-- The common parameter `-Style` is now an `OrderedHashtable`, which allows the declaration order to be preserved.
+- Added the `New-ButtonElement`, `New-QElement`, `New-SlotElement` and `New-SourceElement` cmdlets.
+- Added the `button`, `q`, `slot` and `source` aliases.
+- The common parameter `-Style` is now an `OrderedHashtable`, allowing the declaration order to be preserved.
 
 ## Version [0.9.0](https://github.com/cedx/html.net/compare/v0.8.0...v0.9.0)
 - Added the `New-OlElement`, `New-OptgroupElement`, `New-OptionElement`, `New-OutputElement` and `New-ProgressElement` cmdlets.